requirements.txt
streamlit==1.31.1
streamlit-authenticator==0.2.3
pyyaml==6.0.1
python-dotenv==1.0.0
loguru==0.7.2
pandas==2.1.4
plotly==5.18.0
openpyxl==3.1.2 